==> RCT/01-Analyses.qmd <==
---
title: "Workplace-Intervention_01-Analyses"
author: "SSK"
date: "today"
format: 
  html:
    toc: TRUE
    embed-resources: TRUE
    self-contained-math: TRUE
editor: visual
editor_options: 
  chunk_output_type: console
---

```{r, include=F}
library(tidyverse)
library(psych)
library(knitr)
```

```{r}
data <- read.csv("dataFull_20240130.csv")
```

# Descriptives

::: panel-tabset
#### Condition

```{r, echo=F}
kable(table(data$Condition), col.names = c("Condition","Freq"))
```

#### School / Division

```{r, echo=F}
kable(table(data$schoolDivision_text), col.names = c("School / Division","Freq"))
```

#### Academic / Staff

```{r, echo=F}
kable(table(data$academic_text), col.names = c("Academic / Staff","Freq"))
```

#### Manager

```{r, echo=F, warning=F}
kable(table(data$manager_text), col.names = c("Manager","Freq"))

ggplot(data, aes(x=managerHowMany)) +
  geom_density() +
  labs(x = "# of Employees Managing",
       y = "Density") +
  theme_bw()
```

#### Gender

```{r, echo=F}
kable(table(data$gender_text), col.names = c("Gender","Freq"))
```

#### Race

```{r, echo=F}
kable(table(data$race_text), col.names = c("Race","Freq"))

kable(table(data$race_urm), col.names = c("URM","Freq"))
```

#### SES

```{r, echo=F, warning=F}
kable(psych::describe(data$income))

ggplot(data, aes(income)) +
  geom_density(bins=20, fill="lightblue", alpha=0.8) +
  theme_minimal() +
  xlim(limits=c(1,10)) +
  labs(x = "Income",
       y = "Respondents")

kable(psych::describe(data$ladder))

ggplot(data, aes(ladder)) +
  geom_density(bins=20, fill="lightblue", alpha=0.8) +
  theme_minimal() +
  xlim(limits=c(1,10)) +
  labs(x = "Ladder",
       y = "Respondents")
```
:::

# High-Level Findings

**Confirmatory**

-   Treatment had a significant effect on **multiple-selves / complexity beliefs**
-   Treatment had a marginal effect on **relationship-prioritization** and **perspective-taking**

**Exploratory**

-   Treatment had a marginal effect on **attributions** for co-worker scenario
-   Treatment had no effect for outcomes for manager scenario

**Session 2**

-   Treatment had a marginal / significant effect on **perspective-taking**
-   Treatment had a significant effect on **beliefs about workplace relationships**
-   Treatment even had significant effect on **perspective-taking in manager scenario**

# Key

| Symbol |     Sig. Level      | p           |
|:------:|:-------------------:|:------------|
|   🟢   |     Significant     | \< 0.05     |
|   ️️🟡   | Marginal / Trending | 0.05 - 0.09 |
|   🔴   |   Non-Significant   | \> 0.10     |

# Session 1

# Confirmatory

## 1. Relationship-prioritization 🟡

*How likely are you to prioritize establishing a positive relationship with every single co-worker in your workplace?*

```{r, echo=F}
t.test(data$manipTreat1 ~ data$Condition)

ggplot(data, aes(x=Condition, y=manipTreat1)) +
  geom_bar(stat="summary", position="dodge", fill="lightblue", color = "grey40") +
  expand_limits(y = c(1, 7)) +
  theme_bw() +
  labs(y = "Prioritization of Positive Workplace Relationships")

summary(lm(manipTreat1 ~ Condition + gender, data))

summary(lm(manipTreat1 ~ Condition + manager_text, data))

summary(lm(manipTreat1 ~ Condition * race_urm, data))
```

## 2. Multiple-selves / Complexity Beliefs 🟢

*How much do you agree with the following statement?: Everyone in the workplace is complex inside and is capable of tapping into different parts of themself.*

```{r, echo=F}
t.test(data$manipTreat2_recode ~ data$Condition)

ggplot(data, aes(x=Condition, y=manipTreat2_recode)) +
  geom_bar(stat="summary", position="dodge", fill="lightblue", color = "grey40") +
  expand_limits(y = c(1, 5)) +
  theme_bw() +
  labs(y = "Multiple-selves / Complexity Beliefs")

summary(lm(manipTreat2_recode ~ Condition + gender, data))

summary(lm(manipTreat2_recode ~ Condition + manager_text, data))

summary(lm(manipTreat2_recode ~ Condition * race_urm, data))
```

## 3. Perspective-taking

```{r, include=F}
perspectiveTaking <- data %>% 
  dplyr::select(IRI_PT_1, IRI_PT_2, IRI_PT_3, IRI_PT_4)

alpha(perspectiveTaking, na.rm=T, check.keys=T)
```

$\alpha = 0.63$

**Per preregistration, will therefore look at items individually**

*Over the next 2 weeks while at work...*

::: panel-tabset
#### a. *if I’m sure I’m right about something, I won’t waste much time listening to other people’s arguments.* 🔴

```{r, echo=F}
t.test(data$IRI_PT_1 ~ data$Condition)

summary(lm(IRI_PT_1 ~ Condition + gender, data))

summary(lm(IRI_PT_1 ~ Condition + manager_text, data))

summary(lm(IRI_PT_1 ~ Condition * race_urm, data))
```

#### b. *I will sometimes try to understand my coworkers better by imagining howthings look from their perspective.* 🔴

```{r, echo=F}
t.test(data$IRI_PT_2 ~ data$Condition)

summary(lm(IRI_PT_2 ~ Condition + gender, data))

summary(lm(IRI_PT_2 ~ Condition + manager_text, data))

summary(lm(IRI_PT_2 ~ Condition * race_urm, data))
```

#### c. *I will believe that there are two sides to every question and try to look at them both.* 🔴

```{r, echo=F}
t.test(data$IRI_PT_3 ~ data$Condition)

summary(lm(IRI_PT_3 ~ Condition + gender, data))

summary(lm(IRI_PT_3 ~ Condition + manager_text, data))

summary(lm(IRI_PT_3 ~ Condition * race_urm, data))
```

#### d. *I will try to look at everybody’s side of a disagreement before I make a decision.* 🟡

```{r, echo=F}
t.test(data$IRI_PT_4 ~ data$Condition)

ggplot(data, aes(x=Condition, y=IRI_PT_4)) +
  geom_bar(stat="summary", position="dodge", fill="lightblue", color = "grey40") + 
  expand_limits(y = c(1, 5)) +
  theme_bw() +
  labs(y = "Perspective-taking")

summary(lm(IRI_PT_4 ~ Condition + gender, data))

summary(lm(IRI_PT_4 ~ Condition + manager_text, data))

summary(lm(IRI_PT_4 ~ Condition * race_urm, data))
```
:::

# Exploratory

## Scenario 1

**You just started a new role as a manager. You regularly try to meet with your new employees both as a team and one-on-one to get to know them and to give them the opportunity to get to know you. In general, you have built a good rapport with many of them and they seem to like your leadership style. However, after a few weeks, you notice that one of your employees has seemed to do the bare minimum with some work duties and doesn’t seem mentally present in meetings. From what you have heard, this employee tended to go above and beyond in their work before you started your new role.**

### Attributions 🟡

```{r, echo=F}
t.test(data$attributions_1 ~ data$Condition)

ggplot(data, aes(x=Condition, y=attributions_1)) +
  geom_bar(stat="summary", position="dodge", fill="lightblue", color = "grey40") +
  scale_y_discrete(limit = c("100% Internal", "Internal > External", "Internal = External", "External > Internal", "100% External")) +
  theme_bw() +
  labs(y = "Attributions")

summary(lm(attributions_1 ~ Condition + gender, data))

summary(lm(attributions_1 ~ Condition + manager_text, data))

summary(lm(attributions_1 ~ Condition * race_urm, data))

data %>% 
  filter(!is.na(race_urm)) %>% 
ggplot(aes(x=Condition, y=attributions_1)) +
  facet_wrap(~race_urm) +
  geom_bar(stat="summary", position="dodge") +
  scale_y_discrete(limit = c("100% Internal", "Internal > External", "Internal = External", "External > Internal", "100% External")) +
  theme_bw()
```

### Perspective-Taking 🔴

```{r}
t.test(data$IRI_PT1_1 ~ data$Condition)

summary(lm(IRI_PT1_1 ~ Condition + gender, data))

summary(lm(IRI_PT1_1 ~ Condition + manager_text, data))

summary(lm(IRI_PT1_1 ~ Condition * race_urm, data))

t.test(data$IRI_PT2_1 ~ data$Condition)

summary(lm(IRI_PT2_1 ~ Condition + gender, data))

summary(lm(IRI_PT2_1 ~ Condition + manager_text, data))

summary(lm(IRI_PT2_1 ~ Condition * race_urm, data))
```

## Scenario 2

**You just started reporting to a new manager. After a few weeks, you notice that the new manager has built rapport with some of your colleagues but hasn’t made much effort to get to know you. The new manager has several impromptu meetings per week and seems to forget to loop you in. Although you have been asked to pick up extra work, the manager doesn’t seem to acknowledge your effort or thank you for doing additional tasks.**

### Attributions 🔴

```{r}
t.test(data$attributions_2 ~ data$Condition)

summary(lm(attributions_2 ~ Condition + gender, data))

summary(lm(attributions_2 ~ Condition + manager_text, data))

summary(lm(attributions_2 ~ Condition * race_urm, data))

data %>% 
  filter(!is.na(race_urm)) %>% 
ggplot(aes(x=Condition, y=attributions_2)) +
  facet_wrap(~race_urm) +
  geom_bar(stat="summary", position="dodge") +
  scale_y_discrete(limit = c("100% Internal", "Internal > External", "Internal = External", "External > Internal", "100% External")) +
  theme_bw()
```

### Perspective-Taking 🔴

```{r}
t.test(data$IRI_PT1_2 ~ data$Condition)

summary(lm(IRI_PT1_2 ~ Condition + gender, data))

summary(lm(IRI_PT1_2 ~ Condition + manager_text, data))

summary(lm(IRI_PT1_2 ~ Condition * race_urm, data))

t.test(data$IRI_PT2_2 ~ data$Condition)

summary(lm(IRI_PT2_2 ~ Condition + gender, data))

summary(lm(IRI_PT2_2 ~ Condition + manager_text, data))

summary(lm(IRI_PT2_2 ~ Condition * race_urm, data))
```

## $\Delta$ Scenario 1 - Scenario 2 🔴

### Attributions

```{r}
t.test(data$attributionsDiff ~ data$Condition)

summary(lm(attributionsDiff ~ Condition + gender, data))

summary(lm(attributionsDiff ~ Condition + manager_text, data))

summary(lm(attributionsDiff ~ Condition * race_urm, data))

data %>% 
  filter(!is.na(race_urm)) %>% 
ggplot(aes(x=Condition, y=attributionsDiff)) +
  facet_wrap(~race_urm) +
  geom_bar(stat="summary", position="dodge") +
  theme_bw()
```

### Perspective-Taking

```{r}
t.test(data$IRI_PT1Diff ~ data$Condition)

summary(lm(IRI_PT1Diff ~ Condition + gender, data))

summary(lm(IRI_PT1Diff ~ Condition + manager_text, data))

summary(lm(IRI_PT1Diff ~ Condition * race_urm, data))

t.test(data$IRI_PT2Diff ~ data$Condition)

summary(lm(IRI_PT2Diff ~ Condition + gender, data))

summary(lm(IRI_PT2Diff ~ Condition + manager_text, data))

summary(lm(IRI_PT2Diff ~ Condition * race_urm, data))
```

# Session 2

::: panel-tabset

```{r, include=F}
session2 <- data %>% 
  filter(!is.na(ResponseId_S2))
```

#### Condition

```{r, echo=F}
kable(table(session2$Condition), col.names = c("Condition","Freq"))
```

#### Race

```{r, echo=F}
kable(table(session2$race_urm), col.names = c("URM","Freq"))
```

#### Gender

```{r, echo=F}
kable(table(session2$gender), col.names = c("Gender","Freq"))
```

#### Manager

```{r, echo=F}
kable(table(session2$manager_text), col.names = c("Manager","Freq"))
```

## Perspective-taking

### a. 🟡

```{r, echo=F}
t.test(data$IRI_PT_1_S2 ~ data$Condition)

data %>% 
  filter(!is.na(Condition)) %>% 
ggplot(aes(x=Condition, y=IRI_PT_1_S2)) +
  geom_bar(stat="summary", position="dodge", fill="lightblue", color = "grey40") +
  expand_limits(y = c(1, 5)) +
  theme_bw() +
  labs(y = "Perspective-Taking")
  
summary(lm(IRI_PT_1_S2 ~ Condition + gender, data))

summary(lm(IRI_PT_1_S2 ~ Condition + manager_text, data))

summary(lm(IRI_PT_1_S2 ~ Condition + race_urm, data))
```

### b. 🟢

```{r, echo=F}
t.test(data$IRI_PT_2_S2 ~ data$Condition)

data %>% 
  filter(!is.na(Condition)) %>% 
ggplot(aes(x=Condition, y=IRI_PT_2_S2)) +
  geom_bar(stat="summary", position="dodge", fill="lightblue", color = "grey40") +
  expand_limits(y = c(1, 5)) +
  theme_bw() +
  labs(y = "Perspective-Taking")

summary(lm(IRI_PT_2_S2 ~ Condition + gender, data))

summary(lm(IRI_PT_2_S2 ~ Condition + manager_text, data))

summary(lm(IRI_PT_2_S2 ~ Condition + race_urm, data))
```

### c. 🔴 / 🟢 with URM as covariate

```{r, echo=F}
t.test(data$IRI_PT_3_S2 ~ data$Condition)

summary(lm(IRI_PT_3_S2 ~ Condition + gender, data))

summary(lm(IRI_PT_3_S2 ~ Condition + manager_text, data))

summary(lm(IRI_PT_3_S2 ~ Condition + race_urm, data))
```

### d. 🔴

```{r, echo=F}
t.test(data$IRI_PT_4_S2 ~ data$Condition)

summary(lm(IRI_PT_4_S2 ~ Condition + gender, data))

summary(lm(IRI_PT_4_S2 ~ Condition + manager_text, data))

summary(lm(IRI_PT_4_S2 ~ Condition + race_urm, data))
```

## Relationships

### a. 🟢

*It is important to maintain a positive relationship with co-workers who are hard to work with.*

```{r, echo=F}
t.test(data$workerRelations1_recode_S2 ~ data$Condition)

data %>% 
  filter(!is.na(Condition)) %>% 
ggplot(aes(x=Condition, y=workerRelations1_recode_S2)) +
  geom_bar(stat="summary", position="dodge") +
  expand_limits(y = c(1, 5)) +
  theme_bw()

summary(lm(workerRelations1_recode_S2 ~ Condition + gender, data))

summary(lm(workerRelations1_recode_S2 ~ Condition + manager_text, data))

summary(lm(workerRelations1_recode_S2 ~ Condition + race_urm, data))
```

### b. 🔴

*There are certain co-workers that I come into conflict with and nothing can be done to change that.*

```{r, echo=F}
t.test(data$workerRelations2_recode_S2 ~ data$Condition)

summary(lm(workerRelations2_recode_S2 ~ Condition + gender, data))

summary(lm(workerRelations2_recode_S2 ~ Condition + manager_text, data))

summary(lm(workerRelations2_recode_S2 ~ Condition + race_urm, data))
```

### c. 🔴 / 🟢 with URM as covariate

*A culture of harmonious working relationships is a top priority for me.*

```{r, echo=F}
t.test(data$workerRelations3_recode_S2 ~ data$Condition)

summary(lm(workerRelations3_recode_S2 ~ Condition + gender, data))

summary(lm(workerRelations3_recode_S2 ~ Condition + manager_text, data))

summary(lm(workerRelations3_recode_S2 ~ Condition + race_urm, data))

data %>% 
  filter(!is.na(race_urm)) %>% 
ggplot(aes(x=Condition, y=workerRelations3_recode_S2)) +
  facet_wrap(~race_urm) +
  geom_bar(stat="summary", position="dodge") +
  expand_limits(y = c(1, 5)) +
  theme_bw()
```

# Scenario

## Attributions 🔴

```{r}
t.test(data$attributions_S2 ~ data$Condition)
```

## Perspective-Taking 🟢

```{r}
t.test(data$IRI_PT1_S2 ~ data$Condition)

t.test(data$IRI_PT2_S2 ~ data$Condition)

data %>% 
  filter(!is.na(Condition)) %>% 
ggplot(aes(x=Condition, y=IRI_PT2_S2)) +
  geom_bar(stat="summary", position="dodge", fill="lightblue", color = "grey40") +
  expand_limits(y = c(1, 5)) +
  theme_bw() +
  labs(y = "Manager Vignette Perspective-Taking")
```

# Belonging 🔴

```{r, include=F}
ggplot(data, aes(x=data$belonging_S2)) +
  geom_density() +
  theme_bw()
```

$\alpha = 0.87$

```{r, echo=F}
t.test(data$belonging_S2 ~ data$Condition)
```

# Sense of Fit 🔴

```{r, include=F}
ggplot(data, aes(x=data$fit_S2)) +
  geom_density() +
  theme_bw()
```

$\alpha = 0.89$

```{r, echo=F}
t.test(data$fit_recode_S2 ~ data$Condition)
```
