==> RCT/02-Final-Analyses.qmd <==
---
title: "Workplace-Intervention_01-Analyses"
author: "SSK"
date: "today"
format: 
  html:
    toc: TRUE
    embed-resources: TRUE
    self-contained-math: TRUE
editor: visual
editor_options: 
  chunk_output_type: console
---

```{r, include=F}
library(tidyverse)
library(psych)
library(knitr)
```

```{r}
data <- read.csv("clean_fulldata_20240828.csv")
```

# Descriptives

::: panel-tabset
#### Condition

```{r, echo=F}
kable(table(data$Condition), col.names = c("Condition", "Freq"))
```

#### School / Division

```{r, echo=F}
kable(table(data$schoolDivision), col.names = c("School / Division","Freq"))
```

#### Academic / Staff

```{r, echo=F}
kable(table(data$academic), col.names = c("Academic / Staff","Freq"))
```

#### Manager

```{r, echo=F, warning=F}
kable(table(data$manager), col.names = c("Manager","Freq"))

ggplot(data, aes(x=managerHowMany)) +
  geom_density() +
  labs(x = "# of Employees Managing",
       y = "Density") +
  theme_bw()
```

#### Gender

```{r, echo=F}
kable(table(data$gender), col.names = c("Gender","Freq"))
```

#### Race

```{r, echo=F}
kable(table(data$race), col.names = c("Race","Freq"))
```

#### SES

```{r, echo=F, warning=F}
kable(table(data$income), col.names = c("Income","Freq"))

kable(psych::describe(data$ladder))

ggplot(data, aes(ladder)) +
  geom_density(bins=20, fill="lightblue", alpha=0.8) +
  theme_minimal() +
  xlim(limits=c(1,10)) +
  labs(x = "Ladder",
       y = "Respondents")
```
:::

# High-Level Findings

**Confirmatory**

-   Treatment had a significant effect on **multiple-selves / complexity beliefs**
-   Treatment had a significant effect on **relationship-prioritization** and **perspective-taking**

**Exploratory**

-   Treatment had a marginal effect on **perspective-taking** for manager scenario

**Session 2**

-   Treatment had a marginal / significant effect on **perspective-taking**
-   Treatment had a significant effect on **beliefs about workplace relationships**
-   Treatment even had significant effect on **perspective-taking in manager scenario**

# Key

| Symbol |     Sig. Level      | p           |
|:------:|:-------------------:|:------------|
|   🟢   |     Significant     | \< 0.05     |
|   ️️🟡   | Marginal / Trending | 0.05 - 0.09 |
|   🔴   |   Non-Significant   | \> 0.10     |

# Session 1

# Confirmatory

## 1. Relationship-prioritization 🟢

*How likely are you to prioritize establishing a positive relationship with every single co-worker in your workplace?*

```{r, echo=F}
summary(lm(manipTreat1 ~ Condition + wave, data))

data %>% 
  filter(!is.na(Condition)) %>% 
ggplot(aes(x=Condition, y=manipTreat1)) +
  geom_bar(stat="summary", position="dodge", fill="lightblue", color = "grey40") +
  expand_limits(y = c(1, 7)) +
  theme_bw() +
  labs(y = "Prioritization of Positive Workplace Relationships")
```

## 2. Multiple-selves / Complexity Beliefs 🟢

*How much do you agree with the following statement?: Everyone in the workplace is complex inside and is capable of tapping into different parts of themself.*

```{r, echo=F}
summary(lm(manipTreat2_recode ~ Condition + wave, data))

data %>% 
  filter(!is.na(Condition)) %>% 
ggplot(aes(x=Condition, y=manipTreat2_recode)) +
  geom_bar(stat="summary", position="dodge", fill="lightblue", color = "grey40") +
  expand_limits(y = c(1, 5)) +
  theme_bw() +
  labs(y = "Multiple-selves / Complexity Beliefs")
```

## 3. Perspective-taking

```{r, include=F}
perspectiveTaking <- data %>% 
  dplyr::select(IRI_PT_1, IRI_PT_2, IRI_PT_3, IRI_PT_4)

alpha(perspectiveTaking, na.rm=T, check.keys=T)
```

$\alpha = 0.62$

**Per preregistration, will therefore look at items individually**

*Over the next 2 weeks while at work...*

::: panel-tabset
#### a. 🔴

*if I’m sure I’m right about something, I won’t waste much time listening to other people’s arguments.*

```{r, echo=F}
summary(lm(IRI_PT_1 ~ Condition + wave, data))
```

#### b. 🔴

*I will sometimes try to understand my coworkers better by imagining howthings look from their perspective.*

```{r, echo=F}
summary(lm(IRI_PT_2 ~ Condition + wave, data))
```

#### c. 🟡

*I will believe that there are two sides to every question and try to look at them both.*

```{r, echo=F}
summary(lm(IRI_PT_3 ~ Condition + wave, data))

data %>% 
  filter(!is.na(Condition)) %>% 
ggplot(aes(x=Condition, y=IRI_PT_3)) +
  geom_bar(stat="summary", position="dodge", fill="lightblue", color = "grey40") +
  expand_limits(y = c(1, 5)) +
  theme_bw() +
  labs(y = "Perspective-taking")
```

#### d. 🟢

*I will try to look at everybody’s side of a disagreement before I make a decision.*

```{r, echo=F}
summary(lm(IRI_PT_4 ~ Condition + wave, data))

data %>% 
  filter(!is.na(Condition)) %>% 
ggplot(aes(x=Condition, y=IRI_PT_4)) +
  geom_bar(stat="summary", position="dodge", fill="lightblue", color = "grey40") + 
  expand_limits(y = c(1, 5)) +
  theme_bw() +
  labs(y = "Perspective-taking")
```
:::

# Exploratory

## Burnout 🟡 w/ URM as covariate
```{r}
summary(lm(burnout ~ Condition * URM + wave, data))

data %>% 
  filter(!is.na(Condition)) %>% 
  filter(!is.na(URM)) %>% 
ggplot(aes(x=Condition, y=burnout)) +
  geom_bar(stat="summary", position="dodge", fill="lightblue", color = "grey40") + 
  facet_wrap(~ URM) +
  expand_limits(y = c(1, 8)) +
  theme_bw() +
  labs(y = "Burnout")
```

## Scenario 1

**You just started a new role as a manager. You regularly try to meet with your new employees both as a team and one-on-one to get to know them and to give them the opportunity to get to know you. In general, you have built a good rapport with many of them and they seem to like your leadership style. However, after a few weeks, you notice that one of your employees has seemed to do the bare minimum with some work duties and doesn’t seem mentally present in meetings. From what you have heard, this employee tended to go above and beyond in their work before you started your new role.**

### Attributions 🔴

```{r, echo=F}
summary(lm(attributions_1 ~ Condition + wave + manager, data))
```

### Perspective-Taking 🔴

```{r}
summary(lm(IRI_PT1_1 ~ Condition + wave + manager, data))

summary(lm(IRI_PT2_1 ~ Condition + wave + manager, data))
```

### Hard to Work With 🔴

**How likely is it that you would find it hard to work with this employee at some point in the next month?**

```{r}
summary(lm(hardToWorkWith_1 ~ Condition + wave + manager, data))
```

### Valued 🟢

**How likely is it that you would feel valued by this employee at some point in the next month?**

```{r}
summary(lm(valued_1_recode ~ Condition + wave + manager, data))

data %>% 
  filter(!is.na(Condition)) %>% 
ggplot(aes(x=Condition, y=valued_1_recode)) +
  geom_bar(stat="summary", position="dodge", fill="lightblue", color = "grey40") + 
  expand_limits(y = c(1, 5)) +
  theme_bw() +
  labs(y = "Likelihood of Feeling Valued")
```

### 360 🔴

**I would be willing to give extensive feedback to this employee in a 360 assessment process.**

```{r}
summary(lm(X360_1 ~ Condition + wave + manager, data))
```

### Complaint 🔴

**How likely is it that you would submit a formal complaint about this employee at some point?**

```{r}
summary(lm(complaint_1 ~ Condition + wave + manager, data))
```

## Scenario 2

**You just started reporting to a new manager. After a few weeks, you notice that the new manager has built rapport with some of your colleagues but hasn’t made much effort to get to know you. The new manager has several impromptu meetings per week and seems to forget to loop you in. Although you have been asked to pick up extra work, the manager doesn’t seem to acknowledge your effort or thank you for doing additional tasks.**

### Attributions 🔴

```{r}
summary(lm(attributions_2 ~ Condition + wave + manager, data))
```

### Perspective-Taking 🟡

```{r}
summary(lm(IRI_PT1_2 ~ Condition + wave + manager, data))

data %>% 
  filter(!is.na(Condition)) %>% 
ggplot(aes(x=Condition, y=IRI_PT1_2)) +
  geom_bar(stat="summary", position="dodge", fill="lightblue", color = "grey40") + 
  expand_limits(y = c(1, 5)) +
  theme_bw() +
  labs(y = "Perspective-Taking")

summary(lm(IRI_PT2_2 ~ Condition + wave + manager, data))
```

### 360 🔴

**I would be willing to give extensive feedback to this employee in a 360 assessment process.**

```{r}
summary(lm(X360_2 ~ Condition + wave + manager, data))
```

### Complaint 🔴

**How likely is it that you would submit a formal complaint about this employee at some point?**

```{r}
summary(lm(complaint_2 ~ Condition + wave + manager, data))
```

### Engagement 🔴

**I would be willing to put in extra effort if this manager asked me to.**

```{r}
summary(lm(engagement ~ Condition + wave + manager, data))
```

### Trust 🔴

**How likely is it that you would trust this manager / the leader of your school, college, division or equivalent?**

```{r}
summary(lm(trust_SOW_2 ~ Condition + wave + manager, data))
summary(lm(trustLeader_SOW_2 ~ Condition + wave + manager, data))
```

## $\Delta$ Scenario 1 - Scenario 2 

### Attributions 🔴

```{r}
summary(lm(attributionsDiff ~ Condition + wave + manager, data))
```

### Perspective-Taking 🔴

```{r}
summary(lm(IRI_PT1Diff ~ Condition + wave + manager, data))

summary(lm(IRI_PT2Diff ~ Condition + wave + manager, data))
```

# Session 2

::: panel-tabset

```{r, include=F}
session2 <- data %>% 
  filter(!is.na(ResponseId_S2))
```

#### Condition

```{r, echo=F}
kable(table(session2$Condition), col.names = c("Condition","Freq"))
```

#### School / Division

```{r, echo=F}
kable(table(session2$schoolDivision), col.names = c("School / Division","Freq"))
```

#### Race

```{r, echo=F}
kable(table(session2$race), col.names = c("Race","Freq"))
```

#### Gender

```{r, echo=F}
kable(table(session2$gender), col.names = c("Gender","Freq"))
```

#### Manager

```{r, echo=F}
kable(table(session2$manager), col.names = c("Manager","Freq"))
```

## Perspective-Taking

### a. 🟢

**Over the last week while at work...**

**If I was sure I was right about something, I didn't waste much time listening to other people’s arguments.**

```{r, echo=F}
summary(lm(IRI_PT_1_S2 ~ Condition + wave, data))

data %>% 
  filter(!is.na(Condition)) %>% 
ggplot(aes(x=Condition, y=IRI_PT_1_S2)) +
  geom_bar(stat="summary", position="dodge", fill="lightblue", color = "grey40") +
  expand_limits(y = c(1, 5)) +
  theme_bw() +
  labs(y = "Perspective-Taking")
```

### b. 🟢

**I sometimes tried to understand my coworkers better by imagining how things look from their perspective.**

```{r, echo=F}
summary(lm(IRI_PT_2_S2 ~ Condition + wave, data))

data %>% 
  filter(!is.na(Condition)) %>% 
ggplot(aes(x=Condition, y=IRI_PT_2_S2)) +
  geom_bar(stat="summary", position="dodge", fill="lightblue", color = "grey40") +
  expand_limits(y = c(1, 5)) +
  theme_bw() +
  labs(y = "Perspective-Taking")
```

### c. 🔴

**I believed that there are two sides to every question and tried to look at them both.**

```{r, echo=F}
summary(lm(IRI_PT_3_S2 ~ Condition + wave, data))
```

### d. 🔴

 *I tried to look at everybody’s side of a disagreement before I make a decision.*

```{r, echo=F}
summary(lm(IRI_PT_4_S2 ~ Condition + wave, data))
```

## Trust 🔴
```{r, echo=F}
summary(lm(trustLeader_S2 ~ Condition + wave, data))
summary(lm(trustManager_S2 ~ Condition + wave, data))
```

## Burnout 🔴
```{r}
summary(lm(burnout_S2 ~ Condition + wave, data))
summary(lm(burnout_S2 ~ Condition * manager + wave, data))
summary(lm(burnout_S2 ~ Condition * URM + wave, data))
summary(lm(burnout_S2 ~ Condition * gender + wave, data))
```

## Relationships

### a. 🔴

*It is important to maintain a positive relationship with co-workers who are hard to work with.*

```{r, echo=F}
summary(lm(workerRelations1_S2 ~ Condition + wave, data))
summary(lm(workerRelations1_S2 ~ Condition * manager + wave, data))
```

### b. 🔴

*There are certain co-workers that I come into conflict with and nothing can be done to change that.*

```{r, echo=F}
summary(lm(workerRelations2_S2 ~ Condition + wave, data))
summary(lm(workerRelations2_S2 ~ Condition * manager + wave, data))
```

### c. 🔴 

*A culture of harmonious working relationships is a top priority for me.*

```{r, echo=F}
summary(lm(workerRelations3_S2 ~ Condition + wave, data))
summary(lm(workerRelations3_S2 ~ Condition * manager + wave, data))

```

# Scenario

## Attributions 🔴

```{r}
summary(lm(attributions_S2 ~ Condition + wave + manager, data))
```

## Perspective-Taking 🟡

```{r}
summary(lm(IRI_PT1_S2 ~ Condition + wave + manager, data))

data %>% 
  filter(!is.na(Condition)) %>% 
  filter(!is.na(manager)) %>% 
ggplot(aes(x=Condition, y=IRI_PT1_S2)) +
  geom_bar(stat="summary", position="dodge", fill="lightblue", color = "grey40") +
  facet_wrap(~ manager) +
  expand_limits(y = c(1, 5)) +
  theme_bw() +
  labs(y = "Perspective-Taking")

summary(lm(IRI_PT2_S2 ~ Condition + wave + manager, data))
```

## 360 🟡

```{r}
summary(lm(X360_S2 ~ Condition + wave + manager, data))

data %>% 
  filter(!is.na(Condition)) %>% 
  filter(!is.na(manager)) %>% 
ggplot(aes(x=Condition, y=X360_S2)) +
  geom_bar(stat="summary", position="dodge", fill="lightblue", color = "grey40") +
  facet_wrap(~ manager) +
  expand_limits(y = c(1, 5)) +
  theme_bw() +
  labs(y = "Likelihood of Giving Feedback in 360")

summary(lm(X360Valence_S2 ~ Condition + wave + manager, data))
```

## Complaint 🔴

```{r}
summary(lm(complaint_S2 ~ Condition + wave + manager, data))
```

## Trust 🔴

```{r}
summary(lm(trustManagerScenario_S2 ~ Condition + wave + manager, data))
```

## Evaluative Concerns 🔴

```{r}
summary(lm(evalConcern1_S2 ~ Condition + wave + manager, data))
summary(lm(evalConcern2_S2 ~ Condition + wave + manager, data))
summary(lm(evalConcern3_S2 ~ Condition + wave + manager, data))
summary(lm(evalConcern4_S2 ~ Condition + wave + manager, data))
summary(lm(evalConcern5_S2 ~ Condition + wave + manager, data))
```

## Engagement 🔴

```{r}
summary(lm(engagement1_S2 ~ Condition + wave + manager, data))
summary(lm(engagement2_S2 ~ Condition + wave + manager, data))
summary(lm(engagment3_S2 ~ Condition + wave + manager, data))
```

## Manager 
### Environment 🔴

```{r}
summary(lm(managerEnvironment_S2 ~ Condition + wave + manager, data))
```

### Mindset 🔴

```{r}
summary(lm(managerMindset1_S2 ~ Condition + wave + manager, data))
summary(lm(managerMindset2_S2 ~ Condition + wave + manager, data))
summary(lm(managerMindset3_S2 ~ Condition + wave + manager, data))
summary(lm(managerMindset4_S2 ~ Condition + wave + manager, data))
summary(lm(managerMindset5_S2 ~ Condition + wave + manager, data))
summary(lm(managerMindset6_S2 ~ Condition + wave + manager, data))
```

# Belonging 

```{r, include=F}
belonging <- data %>% 
  dplyr::select(belonging1_S2, belonging2_S2, belonging3_S2, belonging4_S2, belonging5_S2)

keys <- list("belonging1_S2", "belonging2_S2", "belonging3_S2", "belonging4_S2", "belonging5_S2")

alpha(belonging, na.rm=T, check.keys=T)

belonging_S2 <- scoreItems(keys, belonging, min = 1, max = 6)
```

$\alpha = 0.91$

```{r, echo=F}
data$belongingAvg_S2 <- ((data$belonging1_S2 + data$belonging2_S2 + data$belonging3_S2 + data$belonging4_S2 + data$belonging5_S2)/3)

summary(lm(belongingAvg_S2 ~ Condition + wave, data))
summary(lm(belongingAvg_S2 ~ Condition * manager + wave, data))
summary(lm(belongingAvg_S2 ~ Condition * URM + wave, data))
summary(lm(belongingAvg_S2 ~ Condition * gender, data))

data %>% 
  filter(!is.na(Condition)) %>% 
  filter(!is.na(gender)) %>% 
ggplot(aes(x=Condition, y=belongingAvg_S2)) +
  geom_bar(stat="summary", position="dodge", fill="lightblue", color = "grey40") +
  facet_wrap(~ gender) +
  expand_limits(y = c(1, 6)) +
  theme_bw() +
  labs(y = "Belonging")
```

# Sense of Fit 

```{r}
fit <- data %>% 
  dplyr::select(fit1_S2, fit2_S2, fit3_S2)

alpha(fit, na.rm=T, check.keys=T)
```

$\alpha = 0.91$

```{r, include=F}
data$fitAvg_S2 <- ((data$fit1_S2 + data$fit2_S2 + data$fit3_S2)/3)

summary(lm(fitAvg_S2 ~ Condition + wave, data))
summary(lm(fitAvg_S2 ~ Condition * manager + wave, data))
summary(lm(fitAvg_S2 ~ Condition * URM + wave, data))
summary(lm(fitAvg_S2 ~ Condition * gender + wave, data))
```

## Resume 🔴
```{r, echo=F}
summary(lm(intelligent_S2 ~ Condition + wave + manager, data))
summary(lm(motivated_S2 ~ Condition + wave + manager, data))
summary(lm(applicant_1_S2 ~ Condition * manager + wave, data))
summary(lm(applicant_2_S2 ~ Condition * manager + wave, data))
summary(lm(applicant_3_S2 ~ Condition * manager + wave, data))
summary(lm(applicant_4_S2 ~ Condition * manager + wave, data))
```
